+++ apps/native-component-list/src/screens/Video/VideoScreen.tsx
@@ -5,6 +5,14 @@ import ComponentListScreen, { ListElement } from '../ComponentListScreen';
 
 export const VideoScreens = [
   {
+    name: 'Audio Options',
+    route: 'video/audio',
+    options: {},
+    getComponent() {
+      return optionalRequire(() => require('./VideoAudioScreen'));
+    },
+  },
+  {
     name: 'DRM',
     route: 'video/drm',
     options: {},